Policy を用いた Tweet 権限の設定

+++ app/Http/Controllers/Api/TweetController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\UpdateTweetRequest;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
 use App\Services\TweetService;
+use Illuminate\Support\Facades\Gate;
 
 
 class TweetController extends Controller
@@ -27,7 +28,9 @@ class TweetController extends Controller
      */
     public function index()
     {
-        // 🔽 編集
+        // 🔽 追加
+        Gate::authorize('viewAny', Tweet::class);
+
         $tweets = $this->tweetService->allTweets();
         return response()->json($tweets);
     }
@@ -37,9 +40,10 @@ class TweetController extends Controller
      */
     public function store(StoreTweetRequest $request)
     {
-        // バリデーションは削除
-        $tweet = $this->tweetService->createTweet($request);
+        // 🔽 追加
+        Gate::authorize('create', Tweet::class);
 
+        $tweet = $this->tweetService->createTweet($request);
         return response()->json($tweet, 201);
     }
 
@@ -48,6 +52,9 @@ class TweetController extends Controller
      */
     public function show(Tweet $tweet)
     {
+        // 🔽 追加
+        Gate::authorize('view', $tweet);
+
         return response()->json($tweet);
     }
 
@@ -56,9 +63,10 @@ class TweetController extends Controller
      */
     public function update(UpdateTweetRequest $request, Tweet $tweet)
     {
-        // バリデーションは削除
-        $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
+        // 🔽 追加
+        Gate::authorize('update', $tweet);
 
+        $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
         return response()->json($updatedTweet);
     }
     /**
@@ -66,7 +74,9 @@ class TweetController extends Controller
      */
     public function destroy(Tweet $tweet)
     {
-        // 🔽 編集
+        // 🔽 追加
+        Gate::authorize('delete', $tweet);
+
         $this->tweetService->deleteTweet($tweet);
         return response()->json(['message' => 'Tweet deleted successfully']);
     }
