+++ app/Http/Controllers/TweetController.php
@@ -8,6 +8,8 @@ use App\Http\Requests\UpdateTweetRequest;
 use App\Models\Tweet;
 use Illuminate\Http\Request;
 use App\Services\TweetService;
+// 🔽 追加
+use Illuminate\Support\Facades\Gate;
 
 
 class TweetController extends Controller
@@ -26,7 +28,9 @@ class TweetController extends Controller
      */
     public function index()
     {
-        // 🔽 編集
+        // 🔽 追加
+        Gate::authorize('viewAny', Tweet::class);
+
         $tweets = $this->tweetService->allTweets();
         return view('tweets.index', compact('tweets'));
     }
@@ -36,6 +40,9 @@ class TweetController extends Controller
      */
     public function create()
     {
+        // 🔽 追加
+        Gate::authorize('create', Tweet::class);
+
         return view('tweets.create');
     }
 
@@ -44,18 +51,21 @@ class TweetController extends Controller
      */
     public function store(StoreTweetRequest $request)
     {
-        // バリデーションは削除
-        $tweet = $this->tweetService->createTweet($request);
+        // 🔽 追加
+        Gate::authorize('create', Tweet::class);
 
+        $tweet = $this->tweetService->createTweet($request);
         return redirect()->route('tweets.index');
     }
-
 
     /**
      * Display the specified resource.
      */
     public function show(Tweet $tweet)
     {
+        // 🔽 追加
+        Gate::authorize('view', $tweet);
+
         return view('tweets.show', compact('tweet'));
     }
 
@@ -64,6 +74,9 @@ class TweetController extends Controller
      */
     public function edit(Tweet $tweet)
     {
+        // 🔽 追加
+        Gate::authorize('update', $tweet);
+
         return view('tweets.edit', compact('tweet'));
     }
 
@@ -72,21 +85,22 @@ class TweetController extends Controller
      */
     public function update(UpdateTweetRequest $request, Tweet $tweet)
     {
-        // バリデーションは削除
-        $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
+        // 🔽 追加
+        Gate::authorize('update', $tweet);
 
+        $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
         return redirect()->route('tweets.show', $tweet);
     }
-
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(Tweet $tweet)
     {
-        // 🔽 編集
-        $this->tweetService->deleteTweet($tweet);
+        // 🔽 追加
+        Gate::authorize('delete', $tweet);
 
+        $this->tweetService->deleteTweet($tweet);
         return redirect()->route('tweets.index');
     }
 }
